xep0012: turn Reply method into a package function
This maintains the idea of passing around stanzas instead of operating
on them. All stanzas in rfc6120 and rfc6121 go without active
methods (only passive ones that retrieve values). A stanza shouldn't
be responsible for writing meaningful data to a connection; a
connection should do that.

Since getting the XEP-enabled connection is a bit too
cumbersome (potential for panic, alternatively error return) we turn
it into a package function instead. That still fits into the
`scope.Function(stanza, ...)` pattern. We either have
`conn.Function(stanza, ...)` for the rfc6120/rfc6121 functions, or
`xepXXX.Function(stanza, ...)` for XEPs.